It finally saves to the DB. However, the rule_id is not set in the children tables. AKA the cascade.ALL is not doing what I think it should be doing.
Just wanted to get this code state saved into git before I trouble shoot
anymore.

+++ src/main/java/com/ar/marketanalyzer/core/securities/models/parents/PersistableEntityInt.java
@@ -28,7 +28,7 @@ public class PersistableEntityInt implements Serializable{
 		this.id = id;
 	}
 	public void setId(Long id) {
-		this.id = (Integer)id.intValue();
+		this.id = id.intValue();
 	}
 	
 }
